#44 minor refactors
